+++ AE_LoginServer/src/com/aionemu/loginserver/dao/GameServersDAO.java
@@ -23,23 +23,26 @@ import com.aionemu.loginserver.GameServerInfo;
 
 /**
  * DAO that manages GameServers
- *
+ * 
  * @author -Nemesiss-
  */
-public abstract  class GameServersDAO implements DAO 
+public abstract class GameServersDAO implements DAO
 {
-    /**
-     * Returns all gameservers from database.
-     * @return all gameservers from database.
-     */
-    public abstract Map<Integer, GameServerInfo> getAllGameServers();
+	/**
+	 * Returns all gameservers from database.
+	 * 
+	 * @return all gameservers from database.
+	 */
+	public abstract Map<Byte, GameServerInfo> getAllGameServers();
 
-    /**
-     * Returns class name that will be uses as unique identifier for all DAO classes
-     * @return class name
-     */
-    @Override
-    public final String getClassName() {
-        return GameServersDAO.class.getName();
-    }
+	/**
+	 * Returns class name that will be uses as unique identifier for all DAO classes
+	 * 
+	 * @return class name
+	 */
+	@Override
+	public final String getClassName()
+	{
+		return GameServersDAO.class.getName();
+	}
 }
